+++ CHANGELOG.md
@@ -13,6 +13,7 @@ All notable changes to this project will be documented in this file.
 - "Load More" capability to pages modal plausible/analytics#480
 - Unique Visitors (last 30 min) as a top stat in realtime view plausible/analytics#500
 - Pinned filter and date selector rows while scrolling plausible/analytics#472
+- Escape keyboard shortcut to clear all filters plausible/analytics#625
 
 ### Changed
 - Use alpine as base image to decrease Docker image size plausible/analytics#353